requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
openai==1.12.0
python-magic==0.4.27
Pillow==10.2.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0 